feat(psdl_anchoring): extract terminology anchoring into package
Fifth module extraction of the three-tier refactor. Unifies on Workbench's
domain-threaded anchor_ref(ref, psdl_domain, unit) as canonical (resolves
lab/drug name collisions via OMOP-domain filtering) — Inspector previously
had the 392-line undirected version. TerminologyAnchor/TerminologyAnchors
models moved into the package; app/models/schemas.py and
app/services/terminology_anchoring.py are now shims.

Imports repointed to psdl_vocab + psdl_vocab_search packages; zero app.*
imports in the package. Flat layout (package-dir={'psdl_anchoring':'.'}).

Verified: app.main boots from backend cwd; uvicorn 200; creatinine/measurement
anchors to LOINC (domain filter working); full suite green (79 passed, 2
pre-existing errors unchanged).
