+++ backend/psdl_anchoring/pyproject.toml
@@ -1,0 +1,29 @@
+[build-system]
+requires = ["setuptools>=68", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "psdl-anchoring"
+version = "0.1.0"
+description = "Terminology anchoring: PSDL semantic refs → OMOP vocabulary."
+readme = "README.md"
+requires-python = ">=3.9"
+license = {text = "MIT"}
+dependencies = [
+    "pydantic>=2.5.0",
+    "psdl-lang>=0.4.0",
+    "psdl-vocab",
+    "psdl-vocab-search",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.4.0",
+]
+
+[tool.setuptools]
+# This pyproject.toml lives INSIDE the package directory (flat layout).
+# Tell setuptools to treat the current directory as the source of the
+# `psdl_anchoring` package — its __init__.py is right next to this file.
+packages = ["psdl_anchoring"]
+package-dir = {"psdl_anchoring" = "."}
